@@ -9,40 +9,47 @@
   },
   "author": "Andrea Ceccarelli",
   "dependencies": {
+    "@material-ui/core": "^4.9.4",
     "fs": "0.0.1-security",
-    "got": "^8.0.2",
-    "material-ui": "^0.20.0",
-    "metascraper": "^3.4.3",
-    "react": "^15.5.0",
-    "react-chrome-redux": "^1.5.0",
-    "react-dom": "^15.5.0",
-    "react-redux": "^4.0.0",
-    "redux": "^3.0.4",
+    "got": "^10.6.0",
+    "material-ui": "^0.20.2",
+    "metascraper": "^5.11.4",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
-    "require-from-string": "^2.0.1",
+    "redux-thunk": "^2.3.0",
+    "require-from-string": "^2.0.2",
     "scrape-metadata": "^2.0.0",
-    "slack": "^10.0.0",
-    "url-metadata": "^2.1.7"
+    "slack": "^11.0.2",
+    "url-metadata": "^2.4.0",
+    "webext-redux": "^2.1.6"
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "^0.28.8",
-    "eslint": "^4.15.0",
-    "eslint-plugin-json": "^1.2.0",
-    "file-loader": "^1.1.6",
-    "gulp": "^3.9.0",
-    "gulp-load-plugins": "^1.2.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-util": "^3.0.7",
-    "img-loader": "^2.0.0",
-    "rimraf": "^2.5.0",
-    "style-loader": "^0.19.1",
-    "url-loader": "^0.6.2",
-    "webpack": "^2.6.0"
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.8.6",
+    "@babel/helper-module-imports": "^7.8.3",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/preset-env": "^7.8.6",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/register": "^7.8.6",
+    "babel-loader": "^8.0.6",
+    "css-loader": "^3.4.2",
+    "eslint": "^6.8.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-json": "^2.1.0",
+    "eslint-plugin-react": "^7.18.3",
+    "file-loader": "^5.1.0",
+    "gulp": "^4.0.2",
+    "gulp-load-plugins": "^2.0.2",
+    "gulp-rename": "^2.0.0",
+    "gulp-util": "^3.0.8",
+    "img-loader": "^3.0.1",
+    "rimraf": "^3.0.2",
+    "style-loader": "^1.1.3",
+    "url-loader": "^3.0.0",
+    "webpack": "^4.41.6"
   }
 }
