@@ -17,14 +17,14 @@ module.exports = {
   },
 
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.(js)?$/,
         loader: 'babel-loader',
         exclude: /(node_modules)/,
         include: path.join(__dirname, 'src'),
         query: {
-          presets: ['es2015', 'react']
+          presets: ['@babel/preset-react']
         }
       }
     ]
