@@ -1,17 +1,17 @@
-import gulp from 'gulp';
-import loadPlugins from 'gulp-load-plugins';
-import webpack from 'webpack';
-import rimraf from 'rimraf';
+const gulp = require('gulp');
+const loadPlugins = require('gulp-load-plugins');
+const webpack = require('webpack');
+const rimraf = require('rimraf');
 
 const plugins = loadPlugins();
 
-import popupWebpackConfig from './popup/webpack.config';
-import backgroundWebpackConfig from './background/webpack.config';
-// import contentWebpackConfig from './content/webpack.config';
+const popupWebpackConfig = require('./popup/webpack.config');
+const backgroundWebpackConfig = require('./background/webpack.config');
+// const contentWebpackConfig require('./content/webpack.config';
 
-gulp.task('popup-js', ['clean'], (cb) => {
+gulp.task('popup-js', (cb) => {
     webpack(popupWebpackConfig, (err, stats) => {
-        if(err) throw new plugins.util.PluginError('webpack', err);
+        if (err) throw new plugins.util.PluginError('webpack', err);
 
         plugins.util.log('[webpack]', stats.toString());
 
@@ -19,9 +19,9 @@ gulp.task('popup-js', ['clean'], (cb) => {
     });
 });
 
-gulp.task('background-js', ['clean'], (cb) => {
+gulp.task('background-js', (cb) => {
     webpack(backgroundWebpackConfig, (err, stats) => {
-        if(err) throw new plugins.util.PluginError('webpack', err);
+        if (err) throw new plugins.util.PluginError('webpack', err);
 
         plugins.util.log('[webpack]', stats.toString());
 
@@ -39,13 +39,13 @@ gulp.task('background-js', ['clean'], (cb) => {
 //   });
 // });
 
-gulp.task('popup-html', ['clean'], () => {
+gulp.task('popup-html', () => {
     return gulp.src('popup/src/index.html')
         .pipe(plugins.rename('popup.html'))
         .pipe(gulp.dest('./build'))
 });
 
-gulp.task('copy-manifest', ['clean'], () => {
+gulp.task('copy-manifest', () => {
     return gulp.src('manifest.json')
         .pipe(gulp.dest('./build'));
 });
@@ -54,14 +54,14 @@ gulp.task('clean', (cb) => {
     rimraf('./build', cb);
 });
 
-gulp.task('build', ['copy-manifest', 'popup-js', 'popup-html', 'background-js']); // , 'content-js'
+gulp.task('build', gulp.series('copy-manifest', 'popup-js', 'popup-html', 'background-js')); // , 'content-js'
 
-gulp.task('watch', ['default'], () => {
+gulp.task('watch', () => {
     gulp.watch('popup/**/*', ['build']);
     // gulp.watch('content/**/*', ['build']);
     gulp.watch('background/**/*', ['build']);
 });
 
-gulp.task('default', ['build']);
+gulp.task('default', gulp.series('build'));
 
 
