@@ -1,22 +1,17 @@
-import React from 'react';
-import {render} from 'react-dom';
+import React from "react";
+import { render } from "react-dom";
 
-import App from './components/app/App';
-import {Store} from 'react-chrome-redux';
-import {Provider} from 'react-redux';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import App from "./components/app/App";
+import { Store } from "webext-redux";
+import { Provider } from "react-redux";
 
 const proxyStore = new Store({
-  portName: 'example'
+  portName: "example"
 });
 
-
-
-
 render(
-    <Provider store={proxyStore}>
-      <MuiThemeProvider>
-        <App />
-      </MuiThemeProvider>
-    </Provider>
-  , document.getElementById('app'));
+  <Provider store={proxyStore}>
+    <App />
+  </Provider>,
+  document.getElementById("app")
+);
