@@ -1,41 +1,36 @@
-const path = require('path');
+const path = require("path");
 
 module.exports = {
-  entry: [
-    './popup/src/scripts/index.js'
-  ],
+  entry: ["./popup/src/scripts/index.js"],
   output: {
-    filename: 'popup.js',
-    path: path.join(__dirname, '../', 'build'),
-    publicPath: '/'
+    filename: "popup.js",
+    path: path.join(__dirname, "../", "build"),
+    publicPath: "/"
   },
   resolve: {
-    extensions: ['.js', '.jsx', '.scss', '.json'],
-    modules: ['node_modules']
+    extensions: [".js", ".jsx", ".scss", ".json"],
+    modules: ["node_modules"]
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.(jsx|js)?$/,
-        loader: 'babel-loader',
+        loader: "babel-loader",
         exclude: /(node_modules)/,
-        include: path.join(__dirname, 'src'),
+        include: path.join(__dirname, "src"),
         query: {
-          presets: ['es2015', 'react']
+          presets: ["@babel/preset-react"]
         }
       },
       {
         test: /\.css$/,
-        use: ['style-loader', 'css-loader'],
+        use: ["style-loader", "css-loader"],
         exclude: /(node_modules)/,
-        include: path.join(__dirname, 'src'),
+        include: path.join(__dirname, "src")
       },
       {
         test: /\.(jpe?g|png|gif|svg)$/i,
-        use: [
-          'url-loader?limit=10000',
-          'img-loader'
-        ]
+        use: ["url-loader?limit=10000", "img-loader"]
       }
     ]
   }
